+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Use Inter font in Dashboard
 - Hover over any place in the graph
+- Click to filter by 1 day
 
 # 1.0.0 - Dec 12, 2025